requirements.txt
streamlit==1.45.1
pydantic==2.11.4
python-dotenv==1.1.0
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy==2.0.1
